+++ src/main/java/com/example/ApiTirage/Services/PostulantServices.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface PostulantServices {
     Postulants Ajout(Postulants postulants);
     List<Postulants> liste();
-    Postulants Modifier(Postulants pays,Integer Id_postulant);
-    String SupprimerbyId(Integer Id_postulant);
+    Postulants Modifier(Postulants pays,Long Id_postulant);
+    String SupprimerbyId(Long Id_postulant);
 
 }
